+++ mobile/android/android-components/components/browser/storage-sync/src/main/java/mozilla/components/browser/storage/sync/PlacesHistoryStorage.kt
@@ -277,4 +277,12 @@ open class PlacesHistoryStorage(
             }
         }
     }
+
+    override suspend fun deleteHistoryMetadata(key: HistoryMetadataKey) {
+        withContext(writeScope.coroutineContext) {
+            handlePlacesExceptions("deleteHistoryMetadata") {
+                places.writer().deleteHistoryMetadata(key.into())
+            }
+        }
+    }
 }